main/database_sql/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pika